% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{complete.chunk}
\alias{complete.chunk}
\title{Complete Chunk Data
A chunk of data with no missing values from the Murray-Darling Basin Rainfall Data}
\format{A data frame with 576 rows and 78 variables}
\description{
\itemize{
  \item X020020. station No.020020
  \item X024501. station No.024501
  \item ...
}
}
\keyword{datasets}

