% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{date.month}
\alias{date.month}
\title{Date month data
Date information for the Murray-Darling Basin rainfall data}
\format{A vector of dates which length is 1200.}
\description{
Date month data
Date information for the Murray-Darling Basin rainfall data
}
\keyword{datasets}

