\name{EpanK}
\alias{EpanK}

\title{
The Epanechnikov kernel function
}
\description{
The Epanechnikov kernel function that can be passed as an option to the Cutoff() function
}
\usage{
EpanK(x)
}

\arguments{
  \item{x}{
an numerical input vector
}
}


\author{
Lingbing Feng
}

\examples{
x <- seq(-2, 2, len = 30)
EpanK(x)
}

