% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Grmse}
\alias{Grmse}
\title{RMSE give imputed data matrix and the true matrix}
\usage{
Grmse(ximp, xtrue)
}
\arguments{
\item{ximp}{the imputed matrix}

\item{xtrue}{the true matrix}
}
\value{
the RMSE
}
\description{
RMSE give imputed data matrix and the true matrix
}
\examples{
data(hqmr.data)
}

