% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{hqmr.data}
\alias{hqmr.data}
\title{Murray-Darling Basin Rainfall Data}
\format{A data frame with 1200 rows and 79 variables}
\description{
A dataset containing rainfall recordings from 78 gauging stations from the
Murray-Darling Basin in Southeastern Australia.
}
\details{
\itemize{
  \item X020020. station No.020020
  \item X024501. station No.024501
  \item ...
  \item date.month month information
}
}
\keyword{datasets}

