% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nmissing}
\alias{nmissing}
\title{Count the number of missing values in a vector or data matrix}
\usage{
nmissing(x)
}
\arguments{
\item{x}{a vector, matrix or data frame}
}
\value{
the number of missing values (denoted by NA)
}
\description{
Count the number of missing values in a vector or data matrix
}
\examples{
data(hqmr.data)
nmissing(hqmr.data)
}

