\name{UnifK}
\alias{UnifK}
\title{
uniform kernel function
}
\description{
this is the uniform kernel function that can be passed to the Cut function as the kernel option
}
\usage{
UnifK(x)
}

\arguments{
  \item{x}{
x is a numeric vector
}
}

\author{
Lingbing Feng
}

\examples{
x <- seq(-1, 1, len = 30)
UnifK(x)
}
