\name{Grmse}
\alias{Grmse}

\title{
Compute RMSE
}
\description{
Compute mean of RMSE given the imputed data and the true data matrix
}
\usage{
Grmse(ximp, xtrue)
}

\arguments{
  \item{ximp}{
 the imputed data matrix
}
  \item{xtrue}{
the true data matrix
}
}
\details{
ximp and xtrue have to be the same dimension and size. 
}
\value{
It returns the RMSE. 
}

\author{
Lingbing Feng
}
\examples{
# To be continued
}
