% Generated by roxygen2 (4.0.0): do not edit by hand
\name{impCV}
\alias{impCV}
\title{Cross-validation for spatio-temporal imputation}
\usage{
impCV(data, FUN = Cut, date.info = TRUE, cfold = 10, rfold = 10, ...)
}
\arguments{
\item{data}{a data matrix with missing values}

\item{FUN}{the imputation function to be evaluated by cross-validation}

\item{date.info}{logical, if date information is provided in the data.}

\item{cfold}{fold size on the columns}

\item{rfold}{fold size on the rows.}

\item{...}{other arguments}
}
\value{
the cross-validated RMSE
}
\description{
Cross-validation for spatio-temporal imputation
}
\examples{
data(hqmr.data)
# the real cross-validation will take some time to finish
# impCV(hqmr.data)
}

